backend/telegram-webhook/requirements.txt
psycopg2-binary==2.9.9
openai==1.54.0
requests==2.31.0
